fmp4mux: Add support for temporary fragment duration overshoot
Adding "fragment-duration-mode" and "max-fragment-duration-diff"
properties to allow temporary fragment duration overshoot, so that
average fragment duration can be as close as configured
"fragment-duration". This can be useful in HLS production
where absolute time information of a fragment may not be
written in manifest, or a case where player expects sequence
numbers of variants and renditions manifests are all aligned.